signal-engine/requirements.txt
python-dotenv
supabase
pandas
pandas-ta
requests